@@ -1,6 +1,6 @@
 module github.com/sebarcode/siam
 
-go 1.16
+go 1.19
 
 replace github.com/ariefdarmawan/flexpg v0.7.0 => github.com/ariefdarmawan/flexpg v0.7.1
 
@@ -11,4 +11,31 @@ require (
 	github.com/google/uuid v1.3.0
 	github.com/sebarcode/codekit v0.1.1
 	github.com/sebarcode/logger v0.1.1
+)
+
+require (
+	git.kanosolution.net/kano/dbflex v1.2.1 // indirect
+	github.com/ariefdarmawan/byter v0.0.1 // indirect
+	github.com/ariefdarmawan/datahub v0.2.2 // indirect
+	github.com/ariefdarmawan/reflector v0.0.3 // indirect
+	github.com/eaciit/toolkit v0.0.0-20210610161449-593d5fadf78e // indirect
+	github.com/fsnotify/fsnotify v1.5.4 // indirect
+	github.com/hashicorp/hcl v1.0.0 // indirect
+	github.com/logrusorgru/aurora v2.0.3+incompatible // indirect
+	github.com/magiconair/properties v1.8.6 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/pelletier/go-toml v1.9.5 // indirect
+	github.com/pelletier/go-toml/v2 v2.0.1 // indirect
+	github.com/spf13/afero v1.8.2 // indirect
+	github.com/spf13/cast v1.5.0 // indirect
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/spf13/viper v1.12.0 // indirect
+	github.com/subosito/gotenv v1.3.0 // indirect
+	golang.org/x/net v0.0.0-20220520000938-2e3eb7b945c2 // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	golang.org/x/text v0.3.7 // indirect
+	gopkg.in/ini.v1 v1.66.4 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
