@@ -7,17 +7,20 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 
 	"github.com/sebarcode/siam"
 )
 
 type store struct {
 	folderPath string
+	lock       *sync.RWMutex
 }
 
 func NewStorage(loc string) *store {
 	s := new(store)
 	s.folderPath = loc
+	s.lock = new(sync.RWMutex)
 	return s
 }
 
@@ -65,7 +68,7 @@ func (s *store) Store(pool *siam.SessionPool) error {
 
 func (s *store) Get(id string) (*siam.Session, error) {
 	locPath := filepath.Join(s.folderPath, id+".json")
-	bs, e := ioutil.ReadFile(locPath)
+	bs, e := os.ReadFile(locPath)
 	if e != nil {
 		return nil, fmt.Errorf("fail read file %s. %s", id, e.Error())
 	}
@@ -79,12 +82,16 @@ func (s *store) Get(id string) (*siam.Session, error) {
 }
 
 func (s *store) Write(sess *siam.Session) error {
+	s.lock.Lock()
+	defer s.lock.Unlock()
+
 	bs, e := json.Marshal(sess)
 	if e != nil {
 		return fmt.Errorf("fail serializing session %s. %s", sess.SessionID, e.Error())
 	}
 
-	e = ioutil.WriteFile(filepath.Join(s.folderPath, sess.SessionID+".json"), bs, 0644)
+	fileName := filepath.Join(s.folderPath, sess.SessionID+".json")
+	e = os.WriteFile(fileName, bs, 0644)
 	if e != nil {
 		return fmt.Errorf("fail write file %s. %s", sess.SessionID, e.Error())
 	}
@@ -92,6 +99,9 @@ func (s *store) Write(sess *siam.Session) error {
 }
 
 func (s *store) Remove(id string) {
+	s.lock.Lock()
+	s.lock.Unlock()
+
 	locPath := filepath.Join(s.folderPath, id+".json")
 	os.Remove(locPath)
 }
